+++ packages/api/src/main/microservice/exceptions/invalid-input-exception.ts
@@ -4,8 +4,8 @@
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import { RpcException } from '@nestjs/microservices'
+import { QueryMessageException } from './query-message-exception.js'
 
-class InvalidInputException extends RpcException {}
+class InvalidInputException extends QueryMessageException {}
 
 export { InvalidInputException }