+++ packages/micromanage-cli/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [3.1.2](https://github.com/carbon-design-system/carbon-platform/compare/micromanage-cli@3.1.1...micromanage-cli@3.1.2) (2022-10-13)
+
+
+### Misc. 🔮
+
+* update package.json repo and bugs links ([3c089cd](https://github.com/carbon-design-system/carbon-platform/commit/3c089cdde1ddde2a3b9f750680755c4253bfcae2))
+
 ### [3.1.1](https://github.com/carbon-design-system/carbon-platform/compare/micromanage-cli@3.1.0...micromanage-cli@3.1.1) (2022-10-06)